Patch RecursiveScriptModule even if metatensor-learn is not available
Someone could be using a model created with metatensor-learn and
exported to TorchScript, and not have metatensor-learn installed
