+++ python/metatensor_torch/metatensor/torch/learn.py
@@ -71,108 +71,3 @@ module = importlib.util.module_from_spec(spec)
 # with the newly created module
 sys.modules[spec.name] = module
 spec.loader.exec_module(module)
-
-
-# Below, we monkey-patch `torch.jit.RecursiveScriptModule._apply` to also handle
-# metatensor data, similar to what's happening in `metatensor.learn.nn.Module._apply`
-
-_metatensor_data_to = module.nn._module._metatensor_data_to
-
-
-def _apply_metatensor(module):
-    if "_mts_helper" in module._buffers:
-        device = module._mts_helper.device
-        dtype = module._mts_helper.dtype
-
-        if isinstance(module, torch.jit.RecursiveScriptModule):
-            # Get the list of attributes for this module.
-            #
-            # Unfortunately `named_attributes` is not available to Python, so we need to
-            # parse the list from a string.
-            attributes = module._c.dump_to_str(code=False, attrs=True, params=False)
-            attributes = _parse_rsm_attributes(attributes)
-
-            # Update the attributes and re-add them to the module with
-            # `_register_attribute` (which does an update when the attribute already
-            # exists.)
-            for name in attributes:
-                value = module._c.getattr(name)
-
-                value, changed = _metatensor_data_to(value, dtype=dtype, device=device)
-                if changed:
-                    typ = _get_torch_type(value)
-                    module._c._register_attribute(name, typ, value)
-
-
-def _get_torch_type(value):
-    if isinstance(value, torch.ScriptObject):
-        return value._type()
-    elif isinstance(value, int):
-        return torch._C.IntType.get()
-    elif isinstance(value, float):
-        return torch._C.FloatType.get()
-    elif isinstance(value, str):
-        return torch._C.StringType.get()
-    elif isinstance(value, bool):
-        return torch._C.BoolType.get()
-    elif isinstance(value, torch.Tensor):
-        return torch._C.TensorType.get()
-    elif isinstance(value, dict):
-        # assume that all keys/values have the same type, TorchScript would enforce it
-        # anyway
-        key, value = next(iter(value.items()))
-        return torch._C.DictType(_get_torch_type(key), _get_torch_type(value))
-    elif isinstance(value, list):
-        # assume that all values have the same type, TorchScript would enforce it
-        # anyway
-        value = next(iter(value))
-        return torch._C.ListType(_get_torch_type(value))
-    elif isinstance(value, tuple):
-        return torch._C.TupleType([_get_torch_type(v) for v in value])
-    else:
-        return None
-
-
-def _parse_rsm_attributes(string):
-    """
-    Parse the output of ``ScriptModule.dump_to_str()`` to extract the list of attributes
-    on a module.
-    """
-    attributes = []
-    in_attributes = False
-    for line in string.splitlines():
-        if not in_attributes:
-            if "attributes {" in line:
-                in_attributes = True
-                continue
-
-        else:
-            splited = line.split("=")
-            if len(splited) == 2:
-                name = splited[0].strip()
-                attributes.append(name)
-            elif line.endswith("}"):
-                # we are done
-                return attributes
-            else:
-                raise RuntimeError(f"failed to parse line in attributes: '{line}'")
-
-    raise RuntimeError(f"failed to parse attributes section in:\n{string}")
-
-
-# Actual monkey-patching
-_original_rsm_apply = torch.jit.RecursiveScriptModule._apply
-
-
-def _new_rsm_apply(self, fn, recurse=True):
-    output = _original_rsm_apply(self, fn, recurse=recurse)
-    _apply_metatensor(self)
-
-    if recurse:
-        for module in self.children():
-            _apply_metatensor(module)
-
-    return output
-
-
-torch.jit.RecursiveScriptModule._apply = _new_rsm_apply
